Improve unit tests and detect overwritten files.
Start to have separate repository files as input for the parser
unit tests to only show a specific feature and not all at once.

Introduces the build log class to detect if a previously generated
file is overwritten later on. In this case an exception is thrown
which aborts the build.